// frontend-workspace/libs/antero-web/types/src/predefined-colors.enum.ts
export enum PredefinedColors {
  AMX_COLOR_DEFAULT = 'AmxDefault',
  AMX_COLOR_01 = 'AmxColor01',
  AMX_COLOR_02 = 'AmxColor02',
  AMX_COLOR_03 = 'AmxColor03',
  AMX_COLOR_04 = 'AmxColor04',
  AMX_COLOR_05 = 'AmxColor05',
  AMX_COLOR_06 = 'AmxColor06',
  AMX_COLOR_07 = 'AmxColor07',
  AMX_COLOR_08 = 'AmxColor08',
  AMX_COLOR_09 = 'AmxColor09',
  AMX_COLOR_10 = 'AmxColor10',
  AMX_COLOR_11 = 'AmxColor11',
  AMX_COLOR_12 = 'AmxColor12',
  AMX_COLOR_13 = 'AmxColor13',
  AMX_COLOR_14 = 'AmxColor14',
  AMX_COLOR_15 = 'AmxColor15',
  AMX_COLOR_16 = 'AmxColor16',
  AMX_COLOR_17 = 'AmxColor17',
  AMX_COLOR_18 = 'AmxColor18',
  AMX_COLOR_19 = 'AmxColor19',
  AMX_COLOR_20 = 'AmxColor20',
  AMX_COLOR_21 = 'AmxColor21',
  AMX_COLOR_22 = 'AmxColor22',
  AMX_COLOR_23 = 'AmxColor23',
  AMX_COLOR_24 = 'AmxColor24',
  AMX_COLOR_25 = 'AmxColor25',
  AMX_COLOR_26 = 'AmxColor26',
  AMX_COLOR_27 = 'AmxColor27',
  AMX_COLOR_28 = 'AmxColor28',
  AMX_COLOR_29 = 'AmxColor29',
  AMX_COLOR_30 = 'AmxColor30',
  AMX_COLOR_31 = 'AmxColor31'
}